backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
include = '\.pyi?$'
line-length = 100

exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.coverage.run]
data_file = ".cov/.coverage"
source = ["editor", "openfoodfacts_taxonomy_parser"]

[tool.coverage.xml]
output = ".cov/coverage.xml"

[tool.isort] # From https://black.readthedocs.io/en/stable/compatible_configs.html#isort
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.mypy]
ignore_missing_imports = true

[tool.poetry]
name = "openfoodfacts-taxonomy-editor"
version = "0.1.0"
description = ""
authors = ["Open Food Facts Team"]
license = "GNU Affero General Public License v3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["all"], version = "^0.104.1" }
uvicorn = { extras = ["standard"], version = "^0.23.2" }
neo4j = "^5.14.0"
openfoodfacts_taxonomy_parser = { path = "../parser", develop = true }
pygithub = "^2.1.1"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
flake8 = "^6.1.0"
isort = "^5.12.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
httpx = "^0.25.1"
